+++ apps/front/todo-nextjs/src/components/container/Button/Button.tsx
@@ -1,12 +1,12 @@
 ﻿import { ComponentProps } from 'react';
 
-import ButtonPresentational from '~/components/Button/ButtonPresentational';
+import ButtonPresentational from '~/components/presentational/Button/Button';
 
 type Props = {
   presentational: ComponentProps<typeof ButtonPresentational>;
 };
 
-export default function ButtonContainer({ presentational }: Props) {
+export default function Button({ presentational }: Props) {
   return (
     <ButtonPresentational
       text={presentational.text}
