--- apps/front/todo-nextjs/src/storybooks/Button/ButtonContainer.stories.tsx
@@ -1,12 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import ButtonContainer from '~/components/Button/ButtonContainer';
-
-const meta: Meta<typeof ButtonContainer> = {
-  title: 'Button/ButtonContainer',
-  component: ButtonContainer,
-  tags: ['autodocs'],
-  // Add your own control here
-};
-export default meta;
-
-type Story = StoryObj<typeof ButtonContainer>;
